Minor i18n changes and add some platformio environs. * rename locale files from xx_XX.h to xx-XX.h * Enable building different langauges for IRrecvDumpV2 via platformio. * IRrecvDumpV2: Add Comments and add back "Raw Timing" as legacy option. * Clean up platformio.ini files.
