+++ examples/IRMQTTServer/platformio.ini
@@ -5,7 +5,9 @@ src_dir = .
 lib_extra_dirs = ../../
 lib_ldf_mode = deep+
 lib_ignore = examples
-build_flags = -DMQTT_MAX_PACKET_SIZE=768
+build_flags = -DMQTT_MAX_PACKET_SIZE=768 -D_IR_LOCALE_=en-AU
+framework = arduino
+platform = espressif8266
 
 [common]
 lib_deps_builtin =
@@ -26,20 +28,14 @@ lib_deps_external =
   https://github.com/tzapu/WiFiManager.git#development
 
 [env:nodemcuv2]
-platform = espressif8266
-framework = arduino
 board = nodemcuv2
 lib_deps = ${common_esp8266.lib_deps_external}
 
 [env:d1_mini]
-platform = espressif8266
-framework = arduino
 board = d1_mini
 lib_deps = ${common_esp8266.lib_deps_external}
 
 [env:d1_mini_no_mqtt]
-platform = espressif8266
-framework = arduino
 board = d1_mini
 build_flags =
   ${env.build_flags}
@@ -48,13 +44,10 @@ lib_deps = ${common_esp8266.lib_deps_external}
 
 [env:esp32dev]
 platform = espressif32
-framework = arduino
 board = esp32dev
 lib_deps = ${common_esp32.lib_deps_external}
 
 [env:esp01_1m]
-platform = espressif8266
-framework = arduino
 board = esp01_1m
 build_flags =
   ${env.build_flags}
