+++ examples/IRrecvDumpV2/IRrecvDumpV2.ino
@@ -9,6 +9,9 @@
  *  https://github.com/crankyoldgit/IRremoteESP8266/wiki#ir-receiving
  *
  * Changes:
+ *   Version 1.0 October, 2019
+ *     - Internationalisation (i18n) support.
+ *     - Stop displaying the legacy raw timing info.
  *   Version 0.5 June, 2019
  *     - Move A/C description to IRac.cpp.
  *   Version 0.4 July, 2018
@@ -93,6 +96,11 @@ const uint8_t kTimeout = 15;
 // from your device. (e.g. Other IR remotes work.)
 // NOTE: Set this value very high to effectively turn off UNKNOWN detection.
 const uint16_t kMinUnknownSize = 12;
+
+// Legacy (No longer supported!)
+//
+// Change to `true` if you miss/need the old "Raw Timing[]" display.
+#define LEGACY_TIMING_INFO false
 // ==================== end of TUNEABLE PARAMETERS ====================
 
 // Use turn on the save buffer feature for more complete capture coverage.
@@ -112,7 +120,7 @@ void setup() {
 #if DECODE_HASH
   // Ignore messages with less than minimum on or off pulses.
   irrecv.setUnknownThreshold(kMinUnknownSize);
-#endif                  // DECODE_HASH
+#endif  // DECODE_HASH
   irrecv.enableIRIn();  // Start the receiver
 }
 
@@ -123,7 +131,7 @@ void loop() {
     // Display a crude timestamp.
     uint32_t now = millis();
     Serial.printf(D_STR_TIMESTAMP " : %06u.%03u\n", now / 1000, now % 1000);
-    // Check if we got an IR message tha was to big for our capture buffer.
+    // Check if we got an IR message that was to big for our capture buffer.
     if (results.overflow)
       Serial.printf(D_WARN_BUFFERFULL "\n", kCaptureBufferSize);
     // Display the library version the message was captured with.
@@ -134,6 +142,11 @@ void loop() {
     String description = IRAcUtils::resultAcToString(&results);
     if (description.length()) Serial.println(D_STR_MESGDESC ": " + description);
     yield();  // Feed the WDT as the text output can take a while to print.
+#if LEGACY_TIMING_INFO
+    // Output legacy RAW timing info of the result.
+    Serial.println(resultToTimingInfo(&results));
+    yield();  // Feed the WDT (again)
+#endif  // LEGACY_TIMING_INFO
     // Output the results as source code
     Serial.println(resultToSourceCode(&results));
     Serial.println();    // Blank line between entries
