+++ examples/IRrecvDumpV2/platformio.ini
@@ -2,17 +2,42 @@
 src_dir = .
 
 [env]
+; Default platform
+platform = espressif8266
+; Default board
+board = nodemcuv2
+framework = arduino
 lib_extra_dirs = ../../
 lib_ldf_mode = deep+
 lib_ignore = examples
 build_flags =
 
 [env:nodemcuv2]
-platform = espressif8266
-framework = arduino
 board = nodemcuv2
+; build_flags = -D_IR_LOCALE_=en-AU
 
 [env:esp32dev]
 platform = espressif32
-framework = arduino
 board = esp32dev
+; build_flags = -D_IR_LOCALE_=en-AU
+
+[env:de-CH]
+build_flags = -D_IR_LOCALE_=de-CH
+
+[env:de-DE]
+build_flags = -D_IR_LOCALE_=de-DE
+
+[env:en-AU]
+build_flags = -D_IR_LOCALE_=en-AU
+
+[env:en-IE]
+build_flags = -D_IR_LOCALE_=en-IE
+
+[env:en-UK]
+build_flags = -D_IR_LOCALE_=en-UK
+
+[env:en-US]
+build_flags = -D_IR_LOCALE_=en-US
+
+[env:es-ES]
+build_flags = -D_IR_LOCALE_=es-ES
