+++ examples/IRsendDemo/platformio.ini
@@ -5,14 +5,13 @@ src_dir = .
 lib_extra_dirs = ../../
 lib_ldf_mode = deep+
 lib_ignore = examples
+framework = arduino
 build_flags =
 
 [env:nodemcuv2]
 platform = espressif8266
-framework = arduino
 board = nodemcuv2
 
 [env:esp32dev]
 platform = espressif32
-framework = arduino
 board = esp32dev
