+++ examples/Web-AC-control/platformio.ini
@@ -5,6 +5,7 @@ src_dir = .
 lib_extra_dirs = ../../
 lib_ldf_mode = deep+
 lib_ignore = examples
+framework = arduino
 build_flags =
 
 [common]
@@ -26,12 +27,10 @@ lib_deps_external =
 
 [env:nodemcuv2]
 platform = espressif8266
-framework = arduino
 board = nodemcuv2
 lib_deps = ${common_esp8266.lib_deps_external}
 
 [env:esp32dev]
 platform = espressif32
-framework = arduino
 board = esp32dev
 lib_deps = ${common_esp32.lib_deps_external}
