+++ platformio.ini
@@ -1,23 +1,21 @@
 [platformio]
+# Default to building IRrecvDumpV2 if not in a specific example directory.
 src_dir = examples/IRrecvDumpV2
 
 [env]
 lib_extra_dirs = .
 lib_ldf_mode = deep+
 lib_ignore = examples
-build_flags =
+framework = arduino
+platform = espressif8266
+build_flags = -D_IR_LOCALE_=en-AU
 
 [env:nodemcuv2]
-platform = espressif8266
-framework = arduino
 board = nodemcuv2
 
 [env:d1_mini]
-platform = espressif8266
-framework = arduino
 board = d1_mini
 
 [env:esp32dev]
 platform = espressif32
-framework = arduino
 board = esp32dev
