+++ src/locale/en-IE.h
@@ -3,6 +3,6 @@
 // This file will override the default values located in `defaults.h`.
 #ifndef LOCALE_EN_IE_H_
 #define LOCALE_EN_IE_H_
-// Nothing should really need to be set here, as en_IE is the same as en_AU,
+// Nothing should really need to be set here, as en-IE is the same as en-AU,
 // which is the default locale/language.
 #endif  // LOCALE_EN_IE_H__
